-MVC Architecture Design
