+++ src/Main.java
@@ -1,7 +1,9 @@
+
+
+import view.register_login.EntryForm;
+
 import java.sql.SQLException;
 import java.text.ParseException;
-import java.util.ArrayList;
-import java.util.concurrent.atomic.AtomicInteger;
 
 
 public class Main {
@@ -20,22 +22,22 @@ public class Main {
 
 
         //Account creation test
-        /*AccountCreation testAccount = new AccountCreation();
+        /*controller.account.AccountCreation testAccount = new controller.account.AccountCreation();
         testAccount.accountCreator("Samet Berkant", "Koca", date, "editkurama@gmail.com");*/
 
-        //Card creation test
-        /*CardCreation testCard = new CardCreation();
+        //model.datamodel.Card creation test
+        /*controller.card.CardCreation testCard = new controller.card.CardCreation();
 
         testCard.studentCardCreation("Teslime","Koca",date);*/
-        /*SystemConnection connection = new SystemConnection();
+        /*model.db_operations.SystemConnection connection = new model.db_operations.SystemConnection();
         connection.loginSystem("5309760723");
 
-         PaymentService payment = new PaymentService();
+         controller.payment.PaymentService payment = new controller.payment.PaymentService();
         payment.payment(200);*/
 
 
-        // User data getting test
-        /*SystemConnection connection = new SystemConnection();
+        // model.datamodel.User data getting test
+        /*model.db_operations.SystemConnection connection = new model.db_operations.SystemConnection();
         connection.loginSystem("5419664523");*/
 
         /*System.out.println("////////// Ticket System //////////");
@@ -45,7 +47,7 @@ public class Main {
         if (entry==1) {
             System.out.println("Enter your email: ");
             String phone = input.next();
-            SystemConnection connection = new SystemConnection();
+            model.db_operations.SystemConnection connection = new model.db_operations.SystemConnection();
             connection.loginSystem(phone);
         }*/
