+++ src/controller/account/IAuthLoginService.java
@@ -1,3 +1,5 @@
+package controller.account;
+
 public interface IAuthLoginService {
     boolean login(String email);
 }
