+++ src/controller/account/IAuthLogoutService.java
@@ -1,3 +1,5 @@
+package controller.account;
+
 public interface IAuthLogoutService {
     public boolean logout();
 }
