+++ src/controller/account/IAuthRegisterService.java
@@ -1,3 +1,5 @@
+package controller.account;
+
 public interface IAuthRegisterService {
     boolean register();
 }
