+++ src/controller/mail/GmailSender.java
@@ -1,3 +1,5 @@
+package controller.mail;
+
 import javax.mail.*;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
