+++ src/controller/mail/RandomDigitGenerator.java
@@ -1,3 +1,5 @@
+package controller.mail;
+
 import java.util.Random;
 
 public class RandomDigitGenerator {
