+++ src/model/datamodel/User.java
@@ -1,3 +1,5 @@
+package model.datamodel;
+
 import java.util.Date;
 
 public class User {
