+++ src/model/db_operations/DatabaseTransactions.java
@@ -1,3 +1,9 @@
+package model.db_operations;
+
+import controller.account.UserInformationService;
+import model.datamodel.Card;
+import model.datamodel.User;
+
 import java.sql.*;
 
 public class DatabaseTransactions implements IDatabaseService {
