+++ src/model/model_enums/CardType.java
@@ -1,3 +1,5 @@
+package model.model_enums;
+
 public enum CardType {
     STANDARD,
     STUDENT,
